pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "envato-oauth"
version = "1.0.0"
description = "Complete OAuth 2.0 implementation for Envato API with FastAPI server"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Envato API OAuth Package", email = "support@example.com"}
]
keywords = ["envato", "oauth", "api", "fastapi", "authentication", "oauth2"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers", 
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10", 
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: FastAPI",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.35.0",
    "requests>=2.31.0", 
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
selenium = ["selenium>=4.15.0", "webdriver-manager>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0", 
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0", 
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
envato-oauth = "oauth_server:main"
envato-auth = "oauth_server:main"

[project.urls]
Homepage = "https://github.com/yourusername/envato-oauth"
Documentation = "https://github.com/yourusername/envato-oauth/docs"
Repository = "https://github.com/yourusername/envato-oauth"
"Bug Reports" = "https://github.com/yourusername/envato-oauth/issues"

[tool.setuptools]
packages = ["envato_oauth"]
py-modules = ["auth", "envato_oauth", "oauth_server"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.json", ".env.example"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
